move DirectionsClient classes to their own file
